making sure that no column of booleans is saved in /trajectories_data in the results files. booleans are converted to uint8. helps compatibility with Mathematica when reading timeseries directly

+++ tierpsy/gui/MWTrackerViewer.py
@@ -1036,7 +1036,8 @@ class MWTrackerViewer_GUI( MarkersDrawer, PlotCommunicator,
 
         #add this column if it does not exist
         if not 'has_skeleton' in self.trajectories_data:
-            self.trajectories_data['has_skeleton'] = self.trajectories_data['skeleton_id'] >= 0
+            self.trajectories_data['has_skeleton'] = (
+                self.trajectories_data['skeleton_id'] >= 0).astype(np.uint8)
 
         self.updateWormIndexTypeMenu()
         self.updateImage()
